split calc_gradient tests into their own test script fix calc_gradient_FD for qoi_indices when format is array
