+++ openmdao/core/test/test_problem.py
@@ -221,213 +221,6 @@ class TestProblem(unittest.TestCase):
         self.assertEqual(root.connections['c2.x'], ('p1.x', None))
         self.assertEqual(len(root.connections), 2)
 
-    def test_calc_gradient_interface_errors(self):
-
-        root = Group()
-        prob = Problem(root=root)
-        root.add('comp', ExecComp('y=x*2.0'))
-
-        try:
-            prob.calc_gradient(['comp.x'], ['comp.y'], mode='junk')
-        except Exception as error:
-            msg = "mode must be 'auto', 'fwd', 'rev', or 'fd'"
-            self.assertEqual(text_type(error), msg)
-        else:
-            self.fail("Error expected")
-
-        try:
-            prob.calc_gradient(['comp.x'], ['comp.y'], return_format='junk')
-        except Exception as error:
-            msg = "return_format must be 'array' or 'dict'"
-            self.assertEqual(text_type(error), msg)
-        else:
-            self.fail("Error expected")
-
-    def test_calc_gradient(self):
-        root = Group()
-        parm = root.add('parm', IndepVarComp('x', np.array([1., 1., 1., 1.])))
-        comp = root.add('comp', RosenSuzuki())
-
-        root.connect('parm.x', 'comp.x')
-
-        prob = Problem(root)
-        prob.setup(check=False)
-        prob.run()
-
-        indep_list = ['parm.x']
-        unknown_list = ['comp.f', 'comp.g']
-
-        # check that calc_gradient returns proper dict value when mode is 'fwd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fwd', return_format='dict')
-        np.testing.assert_almost_equal(J['comp.f']['parm.x'], np.array([
-            [ -3., -3., -17.,  9.],
-        ]))
-        np.testing.assert_almost_equal(J['comp.g']['parm.x'], np.array([
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]))
-
-        # check that calc_gradient returns proper array value when mode is 'fwd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fwd', return_format='array')
-        np.testing.assert_almost_equal(J, np.array([
-            [-3.,  -3., -17.,  9.],
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]))
-
-        # check that calc_gradient returns proper dict value when mode is 'rev'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='rev', return_format='dict')
-        np.testing.assert_almost_equal(J['comp.f']['parm.x'], np.array([
-            [ -3., -3., -17.,  9.],
-        ]))
-        np.testing.assert_almost_equal(J['comp.g']['parm.x'], np.array([
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]))
-
-        # check that calc_gradient returns proper array value when mode is 'rev'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='rev', return_format='array')
-        np.testing.assert_almost_equal(J, np.array([
-            [-3.,  -3., -17.,  9.],
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]))
-
-        # check that calc_gradient returns proper dict value when mode is 'fd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fd', return_format='dict')
-        np.testing.assert_almost_equal(J['comp.f']['parm.x'], np.array([
-            [ -3., -3., -17.,  9.],
-        ]), decimal=5)
-        np.testing.assert_almost_equal(J['comp.g']['parm.x'], np.array([
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]), decimal=5)
-
-        # check that calc_gradient returns proper array value when mode is 'fd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fd', return_format='array')
-        np.testing.assert_almost_equal(J, np.array([
-            [-3.,  -3., -17.,  9.],
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]), decimal=5)
-
-    def test_calc_gradient_with_indices(self):
-        root = Group()
-        parm = root.add('parm', IndepVarComp('x', np.array([1., 1., 1., 1., 2.])))
-        comp = root.add('comp', RosenSuzuki())
-
-        root.connect('parm.x', 'comp.x', src_indices=[0,1,2,3])
-
-        prob = Problem(root)
-        prob.driver.add_desvar('parm.x', indices=[0,1,2,3])
-        prob.setup(check=False)
-        prob.run()
-
-        indep_list = ['parm.x']
-        unknown_list = ['comp.f', 'comp.g']
-
-        # check that calc_gradient returns proper dict value when mode is 'fwd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fwd', return_format='dict')
-        np.testing.assert_almost_equal(J['comp.f']['parm.x'], np.array([
-            [ -3., -3., -17.,  9.],
-        ]))
-        np.testing.assert_almost_equal(J['comp.g']['parm.x'], np.array([
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]))
-
-        # check that calc_gradient returns proper array value when mode is 'fwd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fwd', return_format='array')
-        np.testing.assert_almost_equal(J, np.array([
-            [-3.,  -3., -17.,  9.],
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]))
-
-        # check that calc_gradient returns proper dict value when mode is 'rev'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='rev', return_format='dict')
-        np.testing.assert_almost_equal(J['comp.f']['parm.x'], np.array([
-            [ -3., -3., -17.,  9.],
-        ]))
-        np.testing.assert_almost_equal(J['comp.g']['parm.x'], np.array([
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]))
-
-        # check that calc_gradient returns proper array value when mode is 'rev'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='rev', return_format='array')
-        np.testing.assert_almost_equal(J, np.array([
-            [-3.,  -3., -17.,  9.],
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]))
-
-        # check that calc_gradient returns proper dict value when mode is 'fd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fd', return_format='dict')
-        np.testing.assert_almost_equal(J['comp.f']['parm.x'], np.array([
-            [ -3., -3., -17.,  9.],
-        ]), decimal=5)
-        np.testing.assert_almost_equal(J['comp.g']['parm.x'], np.array([
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]), decimal=5)
-
-        # check that calc_gradient returns proper array value when mode is 'fd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fd', return_format='array')
-        np.testing.assert_almost_equal(J, np.array([
-            [-3.,  -3., -17.,  9.],
-            [ 3.,   1.,   3.,  1.],
-            [ 1.,   4.,   2.,  3.],
-            [ 6.,   1.,   2., -1.],
-        ]), decimal=5)
-
-    def test_calc_gradient_multiple_params(self):
-        prob = Problem()
-        prob.root = FanIn()
-        prob.setup(check=False)
-        prob.run()
-
-        indep_list   = ['p1.x1', 'p2.x2']
-        unknown_list = ['comp3.y']
-
-        # check that calc_gradient returns proper dict value when mode is 'fwd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fwd', return_format='dict')
-        np.testing.assert_almost_equal(J['comp3.y']['p2.x2'], np.array([[ 35.]]))
-        np.testing.assert_almost_equal(J['comp3.y']['p1.x1'], np.array([[ -6.]]))
-
-        # check that calc_gradient returns proper array value when mode is 'fwd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fwd', return_format='array')
-        np.testing.assert_almost_equal(J, np.array([[-6., 35.]]))
-
-        # check that calc_gradient returns proper dict value when mode is 'rev'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='rev', return_format='dict')
-        np.testing.assert_almost_equal(J['comp3.y']['p2.x2'], np.array([[ 35.]]))
-        np.testing.assert_almost_equal(J['comp3.y']['p1.x1'], np.array([[ -6.]]))
-
-        # check that calc_gradient returns proper array value when mode is 'rev'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='rev', return_format='array')
-        np.testing.assert_almost_equal(J, np.array([[-6., 35.]]))
-
-        # check that calc_gradient returns proper dict value when mode is 'fd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fd', return_format='dict')
-        np.testing.assert_almost_equal(J['comp3.y']['p2.x2'], np.array([[ 35.]]))
-        np.testing.assert_almost_equal(J['comp3.y']['p1.x1'], np.array([[ -6.]]))
-
-        # check that calc_gradient returns proper array value when mode is 'fd'
-        J = prob.calc_gradient(indep_list, unknown_list, mode='fd', return_format='array')
-        np.testing.assert_almost_equal(J, np.array([[-6., 35.]]))
-
     def test_explicit_connection_errors(self):
         class A(Component):
             def __init__(self):
